[refactor] : 클라이언트 및 서버 컴포넌트 단일 진입점 파일 사용으로 구조 개선
