+++ app/dashboard/_components/ProfileDropdownInfo.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 
-import { createClient } from "@/libs/api/supabase";
+import { createClient } from "@/libs/api/client";
 
 interface ProfileInfo {
   nickname: string;
