+++ stores/chatStore2.ts
@@ -1,7 +1,7 @@
 'use client'
 
 import { create } from 'zustand'
-import { createClient } from '@/libs/api/supabase'
+import { createClient } from '@/libs/api/client'
 
 import { Database } from '@/types/database.types'
