+++ stores/notificationStore.ts
@@ -2,7 +2,7 @@
 'use client'
 
 import { create } from 'zustand'
-import { createClient } from '@/libs/api/supabase'
+import { createClient } from '@/libs/api/client'
 
 import { Database } from '@/types/database.types'
